Ticket #48048 - Fix coverity issues - 2015/2/24
Description: Due to the changes made on NSPRIu64 and NSPRI64 along with
this fix:
> Coverity defect 13075 - Explicit null dereferenced (FORWARD_NULL)
> NSPRIu64 and NSPRI64 is defined based upon the architecture and compiler.
casting PRUint64 to (long long unsigned int) in the log print is no longer
needed.  Got rid of them.

Reviewed and advised by rmeggins@redhat.com (Thank you, Rich!!)

https://fedorahosted.org/389/ticket/48048

+++ ldap/servers/plugins/replication/repl5_init.c
@@ -227,7 +227,7 @@ get_repl_session_id (Slapi_PBlock *pb, char *idstr, CSN **csn)
 		if (opid) {
 			slapi_pblock_get (pb, SLAPI_CONN_ID, &connid);
 			PR_snprintf (idstr, REPL_SESSION_ID_SIZE, "conn=%" NSPRIu64 " op=%d",
-					(long long unsigned int)connid, opid);
+					connid, opid);
 		}
 
 		slapi_pblock_get ( pb, SLAPI_OPERATION, &op );
