+++ ldap/servers/plugins/usn/usn.c
@@ -350,7 +350,7 @@ _usn_add_next_usn(Slapi_Entry *e, Slapi_Backend *be)
 
     /* add next USN to the entry; "be" contains the usn counter */
     usn_berval.bv_val = slapi_ch_smprintf("%" NSPRIu64, 
-    		(long long unsigned int)slapi_counter_get_value(be->be_usn_counter));
+                                          slapi_counter_get_value(be->be_usn_counter));
     usn_berval.bv_len = strlen(usn_berval.bv_val);
     slapi_entry_attr_find(e, SLAPI_ATTR_ENTRYUSN, &attr);
     if (NULL == attr) { /* ENTRYUSN does not exist; add it */
@@ -390,7 +390,7 @@ _usn_mod_next_usn(LDAPMod ***mods, Slapi_Backend *be)
     /* add next USN to the mods; "be" contains the usn counter */
     usn_berval.bv_val = counter_buf;
     PR_snprintf(usn_berval.bv_val, USN_COUNTER_BUF_LEN, "%" NSPRIu64, 
-    		(long long unsigned int)slapi_counter_get_value(be->be_usn_counter));
+                slapi_counter_get_value(be->be_usn_counter));
     usn_berval.bv_len = strlen(usn_berval.bv_val);
     bvals[0] = &usn_berval;
     bvals[1] = NULL;
@@ -528,7 +528,7 @@ usn_bepostop (Slapi_PBlock *pb)
     /* if op is not successful, don't increment the counter */
     slapi_pblock_get(pb, SLAPI_RESULT_CODE, &rc);
     if (LDAP_SUCCESS != rc) {
-	/* no plugin failure */
+        /* no plugin failure */
         rc = SLAPI_PLUGIN_SUCCESS;
         goto bail;
     }
@@ -566,7 +566,7 @@ usn_bepostop_modify (Slapi_PBlock *pb)
     /* if op is not successful, don't increment the counter */
     slapi_pblock_get(pb, SLAPI_RESULT_CODE, &rc);
     if (LDAP_SUCCESS != rc) {
-	/* no plugin failure */
+        /* no plugin failure */
         rc = SLAPI_PLUGIN_SUCCESS;
         goto bail;
     }
@@ -617,7 +617,7 @@ usn_bepostop_delete (Slapi_PBlock *pb)
     /* if op is not successful, don't increment the counter */
     slapi_pblock_get(pb, SLAPI_RESULT_CODE, &rc);
     if (LDAP_SUCCESS != rc) {
-	/* no plugin failure */
+        /* no plugin failure */
         rc = SLAPI_PLUGIN_SUCCESS;
         goto bail;
     }
@@ -711,7 +711,7 @@ usn_rootdse_search(Slapi_PBlock *pb, Slapi_Entry* e, Slapi_Entry* entryAfter,
              * then minus 1 from it (except if be_usn_counter has value 0) */
             if (slapi_counter_get_value(be->be_usn_counter)) {
                 PR_snprintf(usn_berval.bv_val, USN_COUNTER_BUF_LEN, "%" NSPRIu64,
-                		(long long unsigned int)slapi_counter_get_value(be->be_usn_counter)-1);
+                            slapi_counter_get_value(be->be_usn_counter)-1);
             } else {
                 PR_snprintf(usn_berval.bv_val, USN_COUNTER_BUF_LEN, "-1");
             }
@@ -734,7 +734,7 @@ usn_rootdse_search(Slapi_PBlock *pb, Slapi_Entry* e, Slapi_Entry* entryAfter,
              * then minus 1 from it (except if be_usn_counter has value 0) */
             if (slapi_counter_get_value(be->be_usn_counter)) {
                 PR_snprintf(usn_berval.bv_val, USN_COUNTER_BUF_LEN, "%" NSPRIu64,
-                		(long long unsigned int)slapi_counter_get_value(be->be_usn_counter)-1);
+                            slapi_counter_get_value(be->be_usn_counter)-1);
             } else {
                 PR_snprintf(usn_berval.bv_val, USN_COUNTER_BUF_LEN, "-1");
             }
